# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "A-ROSA-JE API",
        "description": "Cool API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/address": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get all addresses",
                "operationId": "9ec6f4000f15f43ab455dc6bdec8a404",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Create an address",
                "operationId": "1b883694408b8de9adbe5859004418e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "zip_code": {
                                        "type": "integer",
                                        "maximum": 9999999999
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created"
                    },
                    "500": {
                        "description": "Failed to create address"
                    }
                }
            }
        },
        "/api/v1/address/{id}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get an address by ID",
                "operationId": "ff4170875a219f507803622ffcd8af04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Update an address by ID",
                "operationId": "37c73f3ebbf29028718a4c785ee7b9d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "zip_code": {
                                        "type": "integer",
                                        "maximum": 9999999999
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated"
                    },
                    "500": {
                        "description": "Failed to update address"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Address"
                ],
                "summary": "Delete an address by ID",
                "operationId": "41ae644f8ba71c5fae7631cc35420477",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted"
                    },
                    "500": {
                        "description": "Failed to delete address"
                    }
                }
            }
        },
...
    },
...
}